refactor: CQRS 패턴을 적용하여 PlanService 책임 분리
