+++ src/main/java/com/example/demo/controller/PlanController.java
@@ -5,7 +5,8 @@ import com.example.demo.domain.user.User;
 import com.example.demo.dto.plan.AddPlanRequest;
 import com.example.demo.dto.plan.PlanResponse;
 import com.example.demo.dto.plan.UpdatePlanRequest;
-import com.example.demo.service.PlanService;
+import com.example.demo.service.PlanCommandService;
+import com.example.demo.service.PlanQueryService;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -27,12 +28,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/plans")
 public class PlanController {
 
-    private final PlanService planService;
+    private final PlanCommandService planCommandService;
+    private final PlanQueryService planQueryService;
 
     @PostMapping
     public ResponseEntity<Void> addPlan(@RequestBody AddPlanRequest request,
         @CurrentUser User user) {
-        planService.addMissionToPlan(request.getMissionId(), user);
+        planCommandService.addMissionToPlan(request.getMissionId(), user);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
@@ -41,19 +43,19 @@ public class PlanController {
         @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date,
         @RequestParam(required = false) Boolean isDone,
         @CurrentUser User user) {
-        return ResponseEntity.ok(planService.getPlans(date, isDone, user));
+        return ResponseEntity.ok(planQueryService.getPlans(date, isDone, user));
     }
 
     @PatchMapping("/{id}")
     public ResponseEntity<Void> updatePlanStatus(@PathVariable Long id,
         @RequestBody UpdatePlanRequest request, @CurrentUser User user) {
-        planService.updatePlanStatus(id, request.getIsDone(), user);
+        planCommandService.updatePlanStatus(id, request.getIsDone(), user);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePlan(@PathVariable Long id, @CurrentUser User user) {
-        planService.deletePlan(id, user);
+        planCommandService.deletePlan(id, user);
         return ResponseEntity.noContent().build();
     }
 }
