+++ src/main/java/com/example/demo/service/PlanCommandService.java
@@ -3,13 +3,10 @@ package com.example.demo.service;
 import com.example.demo.domain.mission.Mission;
 import com.example.demo.domain.mission.UserMission;
 import com.example.demo.domain.user.User;
-import com.example.demo.dto.plan.PlanResponse;
 import com.example.demo.repository.MissionRepository;
 import com.example.demo.repository.UserMissionRepository;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class PlanService {
+public class PlanCommandService {
 
     private final UserMissionRepository userMissionRepository;
     private final MissionRepository missionRepository;
@@ -27,7 +24,7 @@ public class PlanService {
         Mission mission = findMissionById(missionId);
         createUserMission(user, mission);
     }
-    
+
     private void validateIfMissionCanBeAdded(Long missionId, User user) {
         LocalDate today = LocalDate.now();
         boolean isNotCompletedMissionExists = userMissionRepository.existsByUserAndMissionIdAndDoneIsFalseAndCreatedAtBetween(
@@ -51,39 +48,26 @@ public class PlanService {
         userMissionRepository.save(userMission);
     }
 
-    @Transactional(readOnly = true)
-    public List<PlanResponse> getPlans(LocalDate date, Boolean isDone, User user) {
-        List<UserMission> userMissions = userMissionRepository.findByUserAndDateAndDone(
-            user,
-            date.atStartOfDay(),
-            date.atTime(LocalTime.MAX),
-            isDone
-        );
-
-        return userMissions.stream()
-            .map(PlanResponse::new)
-            .collect(Collectors.toList());
-    }
-
     public void updatePlanStatus(Long planId, boolean isDone, User user) {
-        UserMission userMission = userMissionRepository.findById(planId)
-            .orElseThrow(() -> new RuntimeException("계획을 찾을 수 없습니다."));
-
-        if (!userMission.getUser().getId().equals(user.getId())) {
-            throw new RuntimeException("본인의 계획만 수정할 수 있습니다.");
-        }
-
+        UserMission userMission = findUserMissionById(planId);
+        validateUserOwnership(userMission, user);
         userMission.updateDone(isDone);
     }
 
     public void deletePlan(Long planId, User user) {
-        UserMission userMission = userMissionRepository.findById(planId)
-            .orElseThrow(() -> new RuntimeException("계획을 찾을 수 없습니다."));
-
-        if (!userMission.getUser().getId().equals(user.getId())) {
-            throw new RuntimeException("본인의 계획만 삭제할 수 있습니다.");
-        }
-
+        UserMission userMission = findUserMissionById(planId);
+        validateUserOwnership(userMission, user);
         userMissionRepository.delete(userMission);
+    }
+
+    private UserMission findUserMissionById(Long planId) {
+        return userMissionRepository.findById(planId)
+            .orElseThrow(() -> new RuntimeException("계획을 찾을 수 없습니다."));
+    }
+
+    private void validateUserOwnership(UserMission userMission, User user) {
+        if (!userMission.getUser().getId().equals(user.getId())) {
+            throw new RuntimeException("본인의 계획만 수정 또는 삭제할 수 있습니다.");
+        }
     }
 }
